+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/Application.java
@@ -3,9 +3,13 @@ package org.linlinjava.litemall.admin;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
 
 @SpringBootApplication(scanBasePackages={"org.linlinjava.litemall.core", "org.linlinjava.litemall.admin","org.linlinjava.litemall.db"})
 @MapperScan("org.linlinjava.litemall.db.dao")
+@EnableTransactionManagement
+@EnableScheduling
 public class Application {
 
     public static void main(String[] args) {
