+++ litemall-db/src/main/java/org/linlinjava/litemall/db/util/OrderUtil.java
@@ -31,6 +31,7 @@ public class OrderUtil {
     public static final Short STATUS_CANCEL = 102;
     public static final Short STATUS_AUTO_CANCEL = 103;
     public static final Short STATUS_REFUND = 202;
+    public static final Short STATUS_REFUND_CONFIRM = 203;
     public static final Short STATUS_AUTO_CONFIRM = 402;
